Added database schema
The schema follows the extracted XML files from the Stack Overflow dataset. The posts table contains the questions and the answers so therefore not all of the columns is used for every row. The post_type decides whether it is an answer or a question. A few columns might be removed in the future. The comments table is pretty straight forward as it only contains a few Ids, text and a foreign key to the related post.