Added script which inserts the XML data
The following script will read two XML files from the same directory as the script and both should follow the Stack Overflow dataset format. The files it will try to load is extractedPosts.xml and extractedComments.xml. The script is definitely not optimal, but it gets the job done. Use the same virtual environment as the django project is using as this script relies on the psycopg2 package for database insertions.