Added "questions" app models
The models does not follow the database.sql schema and the database file might be removed in the future as it might not be needed. These models is more accurate when it comes to null values and whatnot when comparing it to the Stack Overflow dataset. The models has been registered in admin.py so that they are available in the admin panel (/admin).