Added search view
This is the search view that will be used for full text search. Currently it just uses the "contains" function to find posts that contain the search words. The search bar is still not usable, this is just the function it will use. One may manually do the search by going to a url of the following format "/search/<search-word>".