Performance improvements for index and search view
The parent_id of the Post model is now a FK to the PK so that it can be
properly used in Djangos ORM. The reason for it not being a FK already
is because of the dataset. The post-preview template is now using a
field in the query instead of calling the post property function for
every result. A few HTML comments has also been converted to DTL
comments in question.html.